#7 Registro Entrada
Creación del componente formulario y se divide
