+++ src/app/SCT/pages/registro-entrada/registro-entrada.component.html
@@ -1,106 +1,6 @@
-
-<div class="row mt-2">
+<div class="row h-100 my-auto mx-auto">
     <p-toast position="bottom-center"></p-toast>
-    <div class="col-9 mx-auto my-auto">
-        <p-card header="Registro Entrada">
-            <form  class="p-fluid" [formGroup]="registroEntradaForm" autocomplete="off">
-
-                <div class="row mb-2">
-                    <div class="col-6">
-                        <label>Cédula</label>
-                        <div class="p-inputgroup">
-                            <input type="text" pInputText
-                                formControlName="CEDULA"
-                                (keyup.enter)="buscar()">
-                            <span class="p-inputgroup-addon"><i class="pi pi-search" (click)="buscar()"></i></span>
-                        </div>
-                        <small class="p-error block" *ngIf="esInvalido('CEDULA')">
-                            <i class="pi pi-exclamation-circle"></i>
-                            Ingrese una cédula válida
-                        </small>
-                    </div>
-                    <div class="col-6">
-                        <label for="float-input">Tipo de Afiliación</label>
-                        <span class="p-float-label">
-                            <input class="form-control shadow disabled" id="float-input" type="text" pInputText 
-                            formControlName="TIPO_CLIENTE"
-                            >
-                        </span>
-                    </div>
-                </div>
-
-                <div class="row mb-2">
-                    <div class="col-4">
-                        <label for="float-input">Nombre</label>
-                        <span class="p-float-label">
-                            <input id="float-input" type="text" pInputText
-                            formControlName="NOMBRE">
-                        </span>
-                        <small class="p-error block" *ngIf="esInvalido('NOMBRE')">
-                            <i class="pi pi-exclamation-circle"></i>
-                            Ingrese el nombre
-                        </small>
-                    </div>
-                    <div class="col-4">
-                        <label for="float-input">Apellido 1</label>
-                        <span class="p-float-label">
-                            <input id="float-input" type="text" pInputText
-                            formControlName="APELLIDO_1">
-                        </span>
-                        <small class="p-error block" *ngIf="esInvalido('APELLIDO_1')">
-                            <i class="pi pi-exclamation-circle"></i>
-                            Ingrese el Apellido 1
-                        </small>
-                    </div>
-                    <div class="col-4">
-                        <label for="float-input">Apellido 2</label>
-                        <span class="p-float-label">
-                            <input id="float-input" type="text" pInputText
-                            formControlName="APELLIDO_2">
-                        </span>
-                    </div>
-                </div>
-
-                <div class="row mb-4">
-                    <div class="col-6">
-                        <label for="float-input">Área</label>
-                        <input id="float-input" type="text" pInputText
-                            formControlName="AREA">
-                        <!-- <p-dropdown formControlName="selectedCity"
-                        formControlName="AREA"></p-dropdown> -->
-                    </div>
-                    <div class="col-6">
-                        <label for="float-input">Motivo de la visita</label>
-                        <span class="p-float-label">
-                            <input id="float-input" type="text" pInputText
-                            formControlName="MOTIVO"
-                            >
-                        </span>
-                    </div>
-                </div>
-
-
-                <div class="row mb-1">
-                    <div class="col">
-                        <div class="row float-end">
-                            <div class="col">
-                                <p-button label="Borrar" styleClass="p-button-danger" icon="pi pi-trash" (click)="borrar()"></p-button>
-                            </div>
-                            <div class="col">
-                                <p-button label="Registrar" styleClass="p-button-warning" icon="pi pi-check" (click)="registrar()"></p-button>
-                                
-                            </div>
-                        </div>
-                    </div>
-                </div>
-
-            </form>
-        </p-card>
-    </div>
-
-<!-- <div class="row h-100 justify-content-center align-items-center">
-    <p-card header="Registro de Entrada">
+    <p-card class="col-9 mx-auto" header="Registro Entrada">
+        <app-formulario-registro-entrada></app-formulario-registro-entrada>
     </p-card>
-</div> -->
-
-
+</div>
